requirements.txt
asgiref==3.8.1
Django==4.2.22
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.14.0
